+++ resources/views/layout/job/type/table.blade.php
@@ -6,6 +6,7 @@
             @include('component.table.title', ['title' => '#', 'column' => 'id', 'sortable' => true, 'class' => 'text-center'])
             @include('component.table.title', ['title' => 'Name', 'column' => 'name', 'sortable' => true])
             @include('component.table.title', ['title' => 'Color', 'column' => 'color', 'sortable' => true])
+            @include('component.table.title', ['title' => 'Misc Info', 'column' => 'misc', 'sortable' => false])
             {{-- Jika can CRUD maka munculkan tombol--}}
             @if(getUserLevel($config['privilege']) >= CAN_CRUD)
                 @include('component.table.title', ['title' => 'Action', 'column' => 'action', 'sortable'=> false, 'class' => 'text-center align-middle'])
@@ -30,6 +31,20 @@
                     <td>{{ $row->name }}</td>
                     <td>
                         <span class="{{ getBadge($row->color) }}">{{ ucwords($row->color) }}</span>
+                    </td>
+                    <td>
+                        @if($row->actual_date || $row->transport)
+                            @if($row->actual_date)
+                                <span><span class="text-muted">Disable</span> :: Actual Date</span><br>
+                            @endif
+                            @if($row->transport)
+                                <div class="d-inline-flex">
+                                    <span class="me-2 text-primary">Transport:</span>{!! getPrice($row->transport) !!}
+                                </div>
+                            @endif
+                        @else
+                            <span class="text-muted">-</span>
+                        @endif
                     </td>
                     {{--                                    Jika can CRUD maka munculkan tombol--}}
                     @if(getUserLevel($config['privilege']) >= CAN_CRUD)